requirements.txt
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
pandas==2.2.2
openpyxl==3.1.2
pypdf==4.2.0
faiss-cpu==1.8.0
python-dotenv==1.0.1
numpy==1.26.4
tiktoken==0.7.0

